Modularize project settings
