+++ hot_exports/celery.py
@@ -1,17 +1,17 @@
 from __future__ import absolute_import
 
-import os
-
 from celery import Celery
-
-# set the default Django settings module for the 'celery' program.
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'hot_exports.settings')
-
 from django.conf import settings
 
-app = Celery('exports',
-             broker='amqp://',
-             backend='amqp://')
+# celery is going to be executed on the command line or via system scripts
+# it's assumed that DJANGO_SETTINGS_MODULE environment variable is set
+
+app = Celery(
+    'exports',
+    broker='amqp://',
+    backend='amqp://'
+)
+
 
 # Using a string here means the worker will not have to
 # pickle the object when using Windows.
